fixed errors and made changes per the reviewer requirements
